플레이어 체력은 이제 double이다.

+++ src/Battleroyale Server/BattleroyaleServer.h
@@ -11,11 +11,11 @@ public:
 
 	virtual const char* GetIdentifier() const;
 
-	void UpdateMessage(int index, int count, double x, double y, int hp, double direction);
+	void UpdateMessage(int index, int count, double x, double y, double hp, double direction);
 	void GetHurt(int dmg);
 	void Die();
 
-	int health;
+	double health;
 	double attack_cooltime;
 	double inv_time;
